Refactor update status lifecycle boundaries
Split update status into a real package with explicit phase policy, strict persisted-state decoding, and cleanup error handling.
