+++ apps/server/vibesensor/use_cases/updates/job_lifecycle.py
@@ -46,11 +46,9 @@ class UpdateJobLifecycleHandler:
         self._tracker.fail("unexpected", f"Unexpected error: {exc}")
         self._logger.exception("update: unexpected error")
 
-    def handle_cancelled_cleanup_error(self, exc: Exception) -> None:
-        self._logger.warning(
-            "Update cleanup interrupted during cancellation",
-            exc_info=(type(exc), exc, exc.__traceback__),
-        )
+    def handle_cleanup_error(self, exc: Exception) -> None:
+        self._tracker.fail("cleanup", f"Cleanup failed: {exc}")
+        self._logger.exception("update: cleanup error")
 
     async def cleanup_after_update(self) -> None:
         tracker = self._tracker
