+++ apps/server/vibesensor/use_cases/updates/status/__init__.py
@@ -1,0 +1,16 @@
+"""Update status boundary: tracking, persistence, transitions, and runtime details."""
+
+from .runtime_details import collect_runtime_details, hash_tree
+from .state_machine import UpdatePhaseStateMachine, UpdatePhaseTransitionError
+from .state_store import DEFAULT_STATE_PATH, UpdateStateStore
+from .tracker import UpdateStatusTracker
+
+__all__ = [
+    "DEFAULT_STATE_PATH",
+    "UpdatePhaseStateMachine",
+    "UpdatePhaseTransitionError",
+    "UpdateStateStore",
+    "UpdateStatusTracker",
+    "collect_runtime_details",
+    "hash_tree",
+]
